Start down path of provisioning droplets
This includes a new crate called psst which will store development
credentials for things like digital ocean api tokens. It will store it
globally, and prompt you to provide it the first time it needs it.

It's not actually _using_ the token yet, but it's getting late.

This required hardcoding "coconut-estate" into the source code, which I
think is fine, but it was kind of nice to avoid it until now.