Create, attach a smaall volume to secrets keeper
The idea is this will be the `--location` where we store the secrets.
Storing it right on the droplet seems like an option as well, except
that it takes away some flexibility. If we ever need to, we can
provision a new droplet and attach the volume to the other droplet and
then destroy the old one.