Add a docker setup for building binaries for prod
Using the same OS (ubuntu) for building as we'll use in production is
helpful because both will have the same libssl version (1.0.0) which
matters at build time, and both of my development machines have 1.1.
This produces binaries that can run in production.

The volume mounting business is meant to make this workflow fast. Cargo
is good about caching stuff that it can reuse on subsequent builds. So
we need to maintain that cache between builds or everything is very
slow.

Separating the target directory out from the main target directory is
just to help make sure that when we use cargo outside of docker (which
we'll do for local development and for running the toolbox) doesn't get
confused by files created by another cargo, with other permissions.
Sometimes docker is weird with permissions, where it changes the
permissions of your files if it writes them from inside the container.