Add a basic website binary crate
It just renders a micro HTML page and loads a static asset.

Later on, the actix-web service will have more routes that represent the
API, and the JS will be a single-page-app.

I think, anyway.

This is just a stake in the ground to build on. There's gonna be a
website as part of all this. There's a lot more to figure out, but that
part feels figured out.