Add a deployment strategy for secrets-keeper
We provision the instance with a dummy service, and the deploy script
overwrites the dummy with the actual binary and restarts the service.

Next step is to generate the SSL cert, install nginx, and update the
service so it starts the service in SSL mode.

The _next_ next steps will be to teach the toolbox how to talk to a
secrets keeper which is behind a bastion.