Generate an ember app with ember cli
```
cd website && npx ember-cli init --verbose true --name website
```

Currently figuring out if this is viable and what my deployment strategy
will be, so this is currently on a branch. But I like the idea of
keeping all this in its own commit indefinitely so I can know what code
was generated automatically and what did I add later.