Move terraform state to Digital Ocean spaces
Making the state remote means I can more easily collaborate with people,
because they can also apply changes. There isn't locking, unfortunately,
but that's fine for now.

In the short term this just means that I can work on this project on
multiple machines without needing to sync the source code via Dropbox,
which has made things a little painful, especially now that I'm building
an ember app which has a huge honking node_modules folder.

The one downside here is that it doesn't seem like Spaces offers object
versioning, so it's not clear that I'll be able to easily recover the
state file if I want to look at an older version of it. That's something
that Dropbox was incidentally providing that I'm now going without.