Extract api from ember website
The ember app is now served directly by nginx rather than going through
nginx _and_ the rust app.

This commit also removes the SSL code path from the API binary. It
turned out not to be necessary. nginx takes care of decrypting the
request and passing it through to the service. Let's encrypt's certbot
tool takes cares of configuring nginx to do that.

Both apps are available on the same server, and nginx routes the request
to the right one based on the host. Like, if you make a request to
api.coconutestate.top, it will get proxied through to the rust service.
If you make a request to www.coconutestate.top, it will be served from
the filesystem. If you make a request to the IP directly, it will go to
the filesystem, because that one is marked as the default.

This commit also adds a dummy endpoint to the api, just to help confirm
that it's actually up and running.

I think it's acceptable to run both of these on the same droplet for
now. It has 3 cores. I'm not sure how the math works if several requests
come in for the web app and also the api around the same time. The api
uses actix-web, which will start up 3 workers because there are 3 cores.
But how does this all work together? The nice thing with the web app is
that users really only need to load the assets once, and then it's all
in memory in the browser as they click around, so there really shouldn't
be _that many_ requests to the web app. Well, later on if there are many
more image and font assets, that could matter...

I left the default nginx config of worker_connections 1024, so I think
nginx will let the connections just keep coming in ...