Build out website a bit with faux content
This includes setting up CORS so that the front-end is _allowed_ to make
requests to the back-end. In both environments (development and
production), the API is separate and so we need to pass in a value for
each backends, and also have a dynamic API host for the frontend.
Thankfully this is all very doable.