Start managing database schema with diesel
I've successfully used this to create a table in production. I needed to
do a little manual work beforehand to get it working, namely creating a
user with a password in the production database.

I've got diesel set up to output a schema file into the api directory,
because one of the next steps is to have the API code actually connect
to the database.

Another of the next steps, which will be necessary for that one as well,
is to figure out how best to run the database in the development
environment.