Fix: interact with production secrets keeper
Similar to how we use a local copy of the diesel CLI on the production
database, we can use an SSH tunnel to interact with the production
database using a local copy of the toolbox CLI. This lets us both read
and write to the production secrets keeper, tunneling requests via the
bastion.

I decided to take this opportunity to reflect on what ports we use for
each development process, because I think it makes sense to have a
convention about what port to use for these SSH tunnels, and I decided
that it should be the local port + 1000. So if the local secrets keeper
is running on port 5002, the SSH tunnel to production will be on port
6002. I tweaked a few of the ports to optimize for rememberability.