Break api code into multiple files
Plannning to start doing more work here, and I think it'll be nice to
not have it basically all be in one big file.

This also renames the diesel schema file, because shortly we'll be
introducing a graphql schema file and I want it to be very clear which
is which.