Update api to use graphql
There's a lot of room for improvement here. I don't need to create the
schema in each request, it should be possible to persist that across
requests. Same goes for the database connection. Juniper has a concept
called an "executor" which gets passed into the request handler, and the
executor has a "context" which I believe we can use to stash things like
a database connection. That might not be right. That might be how we can
store things like an identifier for the current user. It's a thing
that'll need figuring out, but not yet.