Remove CLI versions from most of the CLIs
I think they're not helpful. I'm not going to bump versions, so they'll
always just stay at 0.1.0. Let's drop that `-V` flag entirely.

In the future if we need to know the version for a particular codebase,
we'll figure it out in a different way. For example, we could update the
deploy script to count the commits and send that number to secrets
keeper, and then it'll get fed into the environment when the app starts.
Not sure how secrets keeper itself will know its version. I'm sure we
can figure that out somehow else though.