Initialize the GraphQL schema just once
Rather than each time a request comes in.

This was just done to flex some rust muscles and try out using the rust
language server's integration with Atom, to help get quick feedback as
writing things, which worked and was pretty cool.

This is fairly inspired by the actix/juniper integrations example:
https://github.com/actix/examples/blob/a677dc8a92003405a8f2331e6126e1f685e5acaa/juniper/src/main.rs