dockerize development environment
The motivation here is just to have more sturdy management of the
various processes needed to do development locally. This doesn't get
much closer toward deploying containers in production, although maybe a
few inches out of a miles long journey.

Managed to make a bit of progress with setting up some shared context in
the graphql api as well, which just kinda arose while working on this
although it doesn't truly belong in this commit.

Worked on this over the last two days or so.

There's a litttle bit of remaining work, some of which I'm aware of and
some that I'm not. What I'm aware of:

- The bin/prod/diesel does not work, at all.
- Not yet sure how to follow SSH tunnels via docker, so bin/prod/diesel
  and bin/toolbox don't really work, but they can still be done outside
  of docker...
- The readme is probably terribly outdated
- Not yet sure how to run the ember tests via docker, because they need
  to run Chrome which isn't available... so just running outside of
  docker for now, which means that