Start to introduce sign-in
This is just on the API level for now. But you can now retrieve a (not
useful or real) token by confirming your email/password.

Next up is to generate some kind of real token, maybe using the JWT
standard, to return to the client, so they can use that to authenticate
more sensitive requests.

Responds with nice errors when the client passes an unknown email or
username, or if the password doesn't match.

Maybe that's more info than I should really provide ... it leaks, for
example, that a matching user was found. A thought for later.