Add sketch of an elm website
Using elm 0.19.0

Ran:

mkdir website
cd website
elm init

---

Start to add some elm tooling

This brings in one of the simple elm examples and some bin scripts to
run it in a development mode, with the feedback provided by reactor, and
also a script to generate a static build that is deployable and
previewable. Still feeling out how to integrate the tooling for both
development and production, but it seems like it could be viable.

---

Sketch out a basic elm SPA

Much to do/learn still.

Some big question marks around things like asset fingerprinting...

Just deployed this to prod though ;)