Add sign-up
When you sign up, you get redirected to the sign-in page, and you still
need to sign-in again. That's a bad user experience, but I think it's
fine enough. It seems to have a glitch where those inputs are
pre-populated, but they really shouldn't be. The app doesn't think they
are, so it thinks they're empty and thus the button is disabled. I'm not
sure what's up with that, but I'm calling it a night.

This commit also factors out some views into their own modules. I'm not
sure if it's really a win because it requires more argument passing of
simiarly-named things, but it at least makes it a little easier to find
things.