Start tracking user email_verified
the idea is that we're collecting emails on sign up, but becuase I
haven't integrated any tool for delivering mail yet, I'm not sending an
email to the email address to verify that the address exists and the
current user does control it. That's a pretty important thing to do
eventually so I'm just putting this little stake in the ground.