Restrict admin page to admins
Previously admins could request the page directly, and then the backend
would reject the requests to read data, but they could still see the
page.

In theory, sufficiently malicious non-admins could still spoof the token
to trick the front end into thinking they're an admin, which would still
allow them to see the page, but then the backend would still reject the
requests for data. Also, they could poke around in the JS assets to try
and figure out what _kinds_ of stuff the admin view has, so this is just
a small win.