Bump a few crates to help compile on linux
Was hitting some errors with finding the local openssl, because on arch
linux I'm using some newer version of openssl, and so I needed to get a
newer version of openssl-sys, which meant getting a newer version of
reqwest.

At the moment, nothing is actually deployed anywhere, but I verified
that this works locally.